Removed potentially flawed code
For a fresh start.
